+++ packages/api-common/source/plugins/pagination/extension.test.ts
@@ -1,8 +1,6 @@
 import { describe } from "@mainsail/test-runner";
 
-import { getConfig } from "./config";
-import { Extension } from "./extension";
-import { pagination } from "./index";
+import { Extension } from "./extension.js";
 
 const fakeH = () => ({ continue: Symbol("continue") });
 
@@ -10,37 +8,7 @@ const makeRoute = (plugins?: object) => ({
 	route: { settings: plugins === undefined ? {} : { plugins } },
 });
 
-describe("pagination/config getConfig", ({ it, assert }) => {
-	it("should return config with default limit applied for valid options", () => {
-		const { config, error } = getConfig({ query: { limit: { default: 100 } } });
-
-		assert.undefined(error);
-		assert.equal(config, { query: { limit: { default: 100 } } });
-	});
-
-	it("should apply default limit when not provided", () => {
-		const { config, error } = getConfig({ query: { limit: {} } });
-
-		assert.undefined(error);
-		assert.equal(config, { query: { limit: { default: 100 } } });
-	});
-
-	it("should return an error for invalid options", () => {
-		const { config, error } = getConfig({ query: { limit: { default: -5 } } });
-
-		assert.undefined(config);
-		assert.defined(error);
-	});
-
-	it("should return an error for non-integer default", () => {
-		const { config, error } = getConfig({ query: { limit: { default: "abc" } } });
-
-		assert.undefined(config);
-		assert.defined(error);
-	});
-});
-
-describe<{ subject: Extension }>("pagination/extension Extension", ({ it, beforeEach, assert }) => {
+describe<{ subject: Extension }>("Extension", ({ it, beforeEach, assert }) => {
 	beforeEach((context) => {
 		context.subject = new Extension({ query: { limit: { default: 100 } } });
 	});
@@ -254,71 +222,5 @@ describe<{ subject: Extension }>("pagination/extension Extension", ({ it, before
 
 		assert.is(request.response.source.meta.existing, "meta-value");
 		assert.is(request.response.source.meta.count, 1);
-	});
-});
-
-describe<{ server: any; extSpy: any }>("pagination/index register", ({ it, beforeEach, assert }) => {
-	beforeEach((context) => {
-		const calls: any[] = [];
-		context.extSpy = calls;
-		context.server = {
-			app: {},
-			ext: (...args: any[]) => calls.push(args),
-		};
-	});
-
-	it("should register onPreHandler and onPostHandler extensions for valid options", ({ server, extSpy }) => {
-		pagination.register(server, { query: { limit: { default: 100 } } });
-
-		assert.is(extSpy.length, 2);
-		assert.is(extSpy[0][0], "onPreHandler");
-		assert.is(extSpy[1][0], "onPostHandler");
-	});
-
-	it("should register with empty options (defaults applied)", ({ server, extSpy }) => {
-		pagination.register(server, {});
-
-		assert.is(extSpy.length, 2);
-	});
-
-	it("should throw the joi error for invalid options", ({ server }) => {
-		assert.throws(() => pagination.register(server, { query: { limit: { default: -1 } } }));
-	});
-
-	it("registered onPreHandler should delegate to the extension", ({ server, extSpy }) => {
-		pagination.register(server, { query: { limit: { default: 100 } } });
-
-		const h = fakeH();
-		const request: any = {
-			...makeRoute({ pagination: { enabled: true } }),
-			query: {},
-		};
-
-		const preHandler = extSpy[0][1];
-		const result = preHandler(request, h);
-
-		assert.is(result, h.continue);
-		assert.is(request.query.page, 1);
-		assert.is(request.query.limit, 100);
-	});
-
-	it("registered onPostHandler should delegate to the extension", ({ server, extSpy }) => {
-		pagination.register(server, { query: { limit: { default: 100 } } });
-
-		const h = fakeH();
-		const request: any = {
-			response: { isBoom: true },
-			...makeRoute({ pagination: { enabled: true } }),
-		};
-
-		const postHandler = extSpy[1][1];
-		const result = postHandler(request, h);
-
-		assert.is(result, h.continue);
-	});
-
-	it("should have expected plugin metadata", () => {
-		assert.is(pagination.name, "hapi-pagination");
-		assert.is(pagination.version, "1.0.0");
 	});
 });
